Update from task 30ec2d9a-4397-4c24-bdb1-3e0a9a3146be

+++ GestorConecta2/app/src/main/java/com/gestorconecta2/ui/components/CommonComponents.kt
@@ -133,7 +133,7 @@ fun PasswordGeneratorDialog(
         if (isAnimating) {
             repeat(20) { iteration ->
                 generatedPassword = generateRandomTemp(passwordLength)
-                delay(timeMillis = 50 + iteration * 5)
+                delay(timeMillis = (50 + iteration * 5).toLong())
             }
             isAnimating = false
         }
